Add GFM support
