+++ readme.md
@@ -1,8 +1,5 @@
 # <img src="source/copy-as-markdown.png" width="45" align="left"> Copy as Markdown
 
-[link-amo]: https://addons.mozilla.org/en-US/firefox/addon/cpy-as-md/
-[link-cws]: https://chrome.google.com/webstore/detail/copy-as-markdown/nlaionblcaejecbkcillglodmmfhjhfi/
-
 > Browser extension to copy hyperlinks, images, and selected text as Markdown
 
 Inspired from [this tweet](https://twitter.com/NicoloRibaudo/status/1143521181196345346) by [@nicolo-ribaudo](https://github.com/nicolo-ribaudo).
@@ -16,27 +13,22 @@ Inspired from [this tweet](https://twitter.com/NicoloRibaudo/status/114352118119
 
 ## Screenshot
 
-![Copy as Markdown](media/screenshot-640x400.png)
+![Copy as Markdown](media/screenshot-1280x800.png)
 
 
 ## Features
 
-- Ability to copy links, images, and selected text as Markdown.
-- Linked images, will have options to individually select link or images.
+The user has access to three actions provided in the context menus with support to handle the following text styles.
 
-When copying links and images, Chrome doesn’t let you extract images alt text or anchors text content to be used in Markdown, instead the links themselves are used as link content. Firefox doesn’t have this limitation though.
+- **Formatted Text**: _Italic_, **bold**, ~~strikethrough~~, and `inline code`.
+- **Unordered and Ordered Lists**: With [task lists](https://github.github.com/gfm/#task-list-items-extension-) support.
+- **Tables**: A [GFM extension](https://github.github.com/gfm/#tables-extension-).
+- **Fenced code blocks**: And language detection using [info strings](https://github.github.com/gfm/#example-112).
+- **Links and Images**
+  - `<a href="https://github.com/notlmn">Laxman</a>` → `[Laxman](https://github.com/notlmn)`
+  - `<img alt="hello world" src="https://notlmn.github.io/image.png">` → `![hello world](https://notlmn.github.io/image.png)`
 
-``` md
-<a href="https://github.com/notlmn">Laxman</a>
-                  ⬇
-[Laxman](https://github.com/notlmn)
-```
-
-``` md
-<img alt="hello world" src="https://notlmn.github.io/image.png">
-                          ⬇
-![hello world](https://notlmn.github.io/image.png)
-```
+  > **Note:** When copying links and images, Chrome doesn’t let you extract images `alt` text or text content of anchor nodes to be used in Markdown, instead the links themselves are used as link content. Firefox doesn’t have this limitation.
 
 
 ### Permissions
@@ -51,3 +43,7 @@ The extension requires the following permission from you for working.
 ## License
 
 [MIT](license)
+
+
+[link-amo]: https://addons.mozilla.org/en-US/firefox/addon/cpy-as-md/
+[link-cws]: https://chrome.google.com/webstore/detail/copy-as-markdown/nlaionblcaejecbkcillglodmmfhjhfi/
