+++ webpack.config.js
@@ -1,7 +1,10 @@
+const fs = require('fs');
 const path = require('path');
+
 const SizePlugin = require('size-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 const TerserPlugin = require('terser-webpack-plugin');
+const DefinePlugin = require('webpack').DefinePlugin;
 
 module.exports = {
 	devtool: 'sourcemap',
@@ -15,6 +18,9 @@ module.exports = {
 	},
 	plugins: [
 		new SizePlugin(),
+		new DefinePlugin({
+			__INJECTIBLE_CODE__: JSON.stringify(fs.readFileSync('./source/injectible-code.js', 'utf-8'))
+		}),
 		new CopyWebpackPlugin([
 			{
 				from: '**/*',
